trying to display thumbnail for video file in cell
