+++ deltachat-ios/View/CustomMessageCell.swift
@@ -45,3 +45,4 @@ open class CustomMessageCell: UICollectionViewCell {
   }
 }
 
+
